Fixed Desktop
Fixed the issue regarding sending the results to the client, instead of
sending all results in a single response to the client, it will now
send smaller chunks on demand. Also filtered on type. This will cause a
longer duration to switch in between types, but this will be necessary
for a mobile version as you don't want to send all results in a single
response as this response might grow to a very large size which is not
optimal on mobile networks.