Removed reusable assets, implemented history api
-Removed reusable assets such as fonts, as they can now be retrieved
from the euscreenxlelements app.

-Added a cached category counts filter, this contains the division over
the items over all the categories, so it can be reused, instead of
calculating it every time a user opens a screen.

-Added a "no-results" and "no-results-found" element to be displayed in
the results element if the user opens the screen or if there are no
results to display.

-Added history API support into the application. Whenever a user makes
a change to his selection (order, filters etc), the URL will
dynamically change to reflect these changes. In this case, when a user
opens an item and clicks back at a later moment, he will still see the
results he saw when he did the search the last time.

-Implemented a loading spinner for the results and the active filters.

-Did some bug fixes and implemented new styles provided by Khris.